Session 2 Assignment - Weekend shopping list fueled by chocolate and alcohol
It had been a hard week. My brain was completely fried from this week’s session. We had had a substitute instructor who proved quite challenging to follow. I figured we could all do with a bit of a party weekend fueled by chocolate and alcohol. My idea was to create a weekend shopping list which basically consisted of what and how many alcoholic beverages we were going to need to get over the class.  
It took me 7 versions in CodePen to get it to work. I also discovered the 'alert' prompt this week, which I preferred, as it was easier for me to work with when testing than the console.

Assignment:
Make your own interactive project using as many of the following as possible
+ Use if...else and conditions using comparison operators (==, ===, !=,
!==, >, <, …) or logical operators (&&, ||, !) to make decisions
+ Use a function to group code
+ Use let to create a variable and then reassign it
+ Share your pen (or pens!) on slack
Bonus?
+ Use functions with one or more return statements
+ Combine multiple logical operators (&&, ||, !) in a condition
+ Investigate how comparison operators work with Strings
+ Research and use a switch case statement
+ Format your code neatly and use helpful variable names