Create Weekend TV Watch List
For this week’s assignment, I wanted to try and include everything I had learnt in sessions 1 & 2, as well as start using Arrays and Loops, plus also ask the user to add their own recommendation and bring that into the final list too. It took me longer than the previous assignments (up to version 9 this time) as I did have a few problems with the functions. But I was very happy with the final result.

Create a program with a list that loops through the items in a list and outputs them in a helpful way.
You should:
+ Use an array to store values
+ Use a loop to iterate through the values in the list
+ Use interpolation and console.log to output the result
+ Use methods
Bonus?
+ Research and use unshift(), shift() and split()
+ Concatenate lists with +
+ Use a while loop
+ Research and use object methods