Create Room Colour Customizer
This assignment was a little more challenging mostly because my idea required the use of an image and that alone caused problems as I don’t have a hosting account and also because I wanted the image to overlay a background that changed colour. 

I could have just changed my idea, but instead I began researching for a way to store images online rather than in a local directory. Initially I found and used a way to store it on my Google Drive, but of course Google are not a hosting site, so I knew it wasn’t a stable option, but it worked as an interim. 

Once I had the rest of my code working, I then went back to researching for a better option. I eventually found a recommendation on Stack Exchange for Cloudinary which has been a life saver for my final project.  

I also tried a version using columns but that proved just a little bit too difficult for me at this time as I wasn’t happy with how one of the images was behaving, so I amended my design and went back to a single column layout centred in the middle of the page.

The other problem with this assignment was the instructor had problems with connection this week and wasn’t able to finish the class. We were actually given a pass with our homework, but I decided to forge ahead and finish the studies myself and submit my assignment anyway.

Create a web app that updates the web page.
You should:
+ Make changes to the HTML and/or CSS using JavaScript
+ Use an event to trigger a change to a web page
Bonus?
+ Add HTML elements to a web page using the DOM
+ Remove HTML elements from a web page using the DOM
+ Research and use other DOM events such as onmouseover
+ Use arrays, loops and/or conditional logic in JavaScript